svn path=/branches/arwinss/; revision=42665

+++ reactos/drivers/input/i8042prt/mouse.c
@@ -638,7 +638,7 @@ i8042MouResetIsr(
 	DeviceExtension->MouseTimeoutState = TimeoutStart;
 	PortDeviceExtension = DeviceExtension->Common.PortDeviceExtension;
 
-	switch (DeviceExtension->MouseResetState)
+	switch ((ULONG)DeviceExtension->MouseResetState)
 	{
 		case 1100: /* the first ack, drop it. */
 			DeviceExtension->MouseResetState = ExpectingReset;
